+++ src/App.js
@@ -8,8 +8,8 @@ import MyDetailPage from "./pages/MyDetailpage/MyDetailPage";
 import UploadPage from "./pages/UploadPage/UploadPage";
 import UploadDonePage from "./pages/UploadPage/UploadDonePage";
 import NotFoundPage from "./pages/NotFoundPage/NotFoundPage";
-import KakaoHandler from "./components/KakaoHandler";
-import InstagramHandler from "./components/InstagramHandler";
+import KakaoHandlerPage from "./pages/KakaoHandlerPage/KakaoHandlerPage";
+import InstagramHandlerPage from "./pages/InstagramHandlerPage/InstagramHandlerPage";
 import AuthLayout from "./layouts/AuthLayout";
 
 function App() {
@@ -27,12 +27,12 @@ function App() {
           <Route path="/profile/setting" element={<SettingPage />}></Route>
           <Route path="/upload" element={<UploadPage />}></Route>
           <Route path="/upload-done" element={<UploadDonePage />}></Route>
-          <Route path="/kakao/callback" element={<KakaoHandler />}></Route>
-          <Route
-            path="/instagram/callback"
-            element={<InstagramHandler />}
-          ></Route>
         </Route>
+        <Route path="/kakao/callback" element={<KakaoHandlerPage />}></Route>
+        <Route
+          path="/instagram/callback"
+          element={<InstagramHandlerPage />}
+        ></Route>
         <Route path="*" element={<NotFoundPage />}></Route>
       </Routes>
       <NavigationBar />
